environment/requirements.txt
flatlib==0.2.3
pyswisseph==2.10.3.2
